dags/requirements.txt
protobuf==3.20.*
apache-airflow-providers-google==7.0.0
apache-airflow-providers-cncf-kubernetes==3.1.2